API

Description

Contains the REST API for Sudoku.

Documentation

type SudokuApi = ("health" :> Get '[JSON] String) :<|> (("grid" :> (Capture "gridId" Integer :> Get '[JSON] Grid)) :<|> (("grid" :> (Capture "gridId" Integer :> (ReqBody '[JSON] Grid :> Put '[JSON] Int64))) :<|> (("grid" :> (Capture "gridId" Integer :> ("solve" :> (ReqBody '[JSON] Grid :> Put '[JSON] Grid)))) :<|> (("cell" :> (Capture "cellId" Integer :> ("reveal" :> (ReqBody '[JSON] Cell :> Put '[JSON] Cell)))) :<|> (("grid" :> (Capture "gridId" Integer :> ("reveal" :> (ReqBody '[JSON] Grid :> Put '[JSON] Grid)))) :<|> ("grid" :> (Capture "gridId" Integer :> ("clear" :> Put '[JSON] Grid))))))))

GET: /health. Returns JSON.

GET: /grid/<id>. Returns JSON Grid for the given id.

PUT: /grid/<id>. For given JSON ReqBody Grid with id, saves the grid and returns JSON Int64 representing number of saved cells + saved grid.

PUT: /grid/<id>/solve, For given JSON ReqBody Grid with id, solves and returns the grid.

PUT: /cell/<id>/reveal, For given JSON ReqBody Cell with id, reveals the cell.

PUT: /grid/<id>/reveal, For given JSON ReqBody Grid with id, reveals the grid.

PUT: /grid/<id>/clear, For given gridId id, clears the grid.

sudokuApi :: Proxy SudokuApi

Returns SudokuApi behind Proxy.