Service

Description

Contains the main web server functions.

Documentation

health :: Handler String

Simple health check. Returns a Handler "Sudoku is up!"

getGridById :: Connection -> Integer -> Handler Grid

For a Connection and Integer gridId, returns a Handler Grid. Throws 500 if an error is encountered.

saveGrid :: Connection -> Integer -> Grid -> Handler Int64

For a Connection, Integer gridId, and Grid, saves the grid and returns the total saved objects (Grid + cells).

solveGrid :: Connection -> Integer -> Grid -> Handler Grid

For a Connection, Integer gridId, and Grid, attempts to solve the grid. If the grid cannot be solved an erro message is printed to the console and 500 is thrown.

revealCell :: Connection -> Integer -> Cell -> Handler Cell

For a Connection, Integer cellId, and Cell, reveals a cell.

revealGrid :: Connection -> Integer -> Grid -> Handler Grid

For a Connection, Integer gridId, and Grid, reveals the grid.

clearGrid :: Connection -> Integer -> Handler Grid

For a Connection, Integer gridId, and Grid, clears the grid.