Service.Internal

Description

Contains internal functions for Sudoku.

Documentation

blankGrid :: Grid

Returns a blank Grid

gridTToGrid :: [GridT] -> Maybe [Cell] -> Maybe Grid

For a [GridT] of size 1, and Just [Cell], returns Just Grid. Otherwise returns Nothing.

cellTToCell :: [CellT] -> Maybe [Cell]

For a [CellT], returns Just [Cell]. Otherwise returns Nothing.

solve :: Grid -> Maybe Grid

If Grid g can be solved returns Just g', where g' is solved. Otherwise returns Nothing.

allGuessesForCell :: Grid -> Cell -> [Grid]

For a grid and cell, returns a list of grids where each new grid is all possible guesses for cell.

getEmptyCell :: [Cell] -> Maybe Cell

Returns an empty cell, if any exists.

getSolvedGrid :: [Maybe Grid] -> Maybe Grid

Returns a solved grid, if any exists.

validate :: Grid -> Bool

Returns True if the Grid is valid by the rules of Sudoku, False otherwise.

validateSection :: [Cell] -> Bool

For a Cell section (i.e., row, column, or box), returns True if the section is valid, False otherwise.

Works by filtering the [Cell] into [[Cell]] where each list contains the list of all cells with realValue \(r \in {1,2,\ldots 9}\). After that we test that each list has length < 2.

revealCell :: Cell -> Cell

Marks a cell as revealed.

revealGrid :: Grid -> Grid

Marks a grid as revealed.

markGridSolved :: Grid -> Grid

Marks a grid as solved.

markCorrectCellsRevealed :: [Cell] -> [Cell]

Marks correct cells as revealed.

revealAllCells :: [Cell] -> [Cell]

Marks all cells as revealed.