Sudoku.Generate

Description

Generate random Sudoku grids.

Documentation

generateSolved :: IO Sudoku

Generate a solved Sudoku grid.

generateSolveable :: Int -> IO Sudoku

Generate a solveable Sudoku grid with a given number of empty spaces

generateSolvedFromFirst :: Pos -> Int -> Sudoku

Given a Sudoku grid with only one number, generate a solved grid.

generateSolvedFromGen :: (Monad m, RandomGen g) => g -> m Sudoku

Generate a solved grid using a RandomGen.