Sudoku.Server

Description

This module lets you run the Sudoku solver as an API. The following paths are supported:

/board (GET)

Query parameters:

blanks: the number of blank spaces desired. Use this parameter to control the difficulty of the board

Returns:

{
  board: [[...]]    # a 9x9 array containing an unsolved Sudoku grid. Blanks are represented by 0.
}

/solve (POST)

JSON body:

{
  board: [[...]]    # a 9x9 array containing an unsolved Sudoku grid.
}

Returns:

{
  board: [[...]]    # a 9x9 array containing the solved Sudoku grid.
}

Documentation

module Sudoku.Server.Get

module Sudoku.Server.Post

module Sudoku.Server.Util

server :: Application

The main function of the Server module.