Sudoku.Solve

Description

Sudoku puzzle solving.

Possible entries of a cell

type Options = OneOrMany Int

Options represents our current knowledge about the potential values of a particular cell in a grid.

Utility functions for Options

countOptions :: Int -> [Options] -> Int

Count the occurrences of an Int in a list of Options

toList :: Options -> [Int]

Convert an Options object to a list of integers

makeOptions :: [Int] -> Options

Convert a list of integers to an Options object

lengthOptions :: Options -> Int

Get the length of an Options object.

deleteOption :: Int -> Options -> Options

Delete an option from an Options object.

Sudoku grid plus current knowledge

type SudokuWithOptions = Grid Options

SudokuWithOptions represents a partially complete sudoku grid, and the knowledge we have about the potential values of each cell

Utility functions for SudokuWithOptions

setOptions :: SudokuWithOptions -> Pos -> [Int] -> SudokuWithOptions

Set the options in a cell of the grid

optionsToNormal :: SudokuWithOptions -> Sudoku

Convert from SudokuWithOptions to Sudoku, only taking cells which have only 1 option

applyUntilStatic :: (SudokuWithOptions -> SudokuWithOptions) -> SudokuWithOptions -> SudokuWithOptions

Given a function that transforms a SudokuWithOptions and a starting grid, repeatedly apply the function until it converges to a fixed point.

Basic solving techniques

improve :: Sudoku -> Sudoku

Improve a Sudoku grid as much as possible using the techniques defined in mBasic.

More advanced solving techniques

backtrack :: Sudoku -> Maybe Sudoku

Use backtracking to try to find one solution of a suokdu grid. Returns Nothing if there is no solution.

backtrackList :: Sudoku -> [Sudoku]

Use backtracking to try to find all solutions of a suokdu grid. Returns Nothing if there is no solution.