Sudoku.Solve.Backtracking

Description

More advanced solution concepts using backtracking techniques.

Documentation

backtrack :: Sudoku -> Maybe Sudoku

Use backtracking to try to find one solution of a suokdu grid. Returns Nothing if there is no solution.

backtrackList :: Sudoku -> [Sudoku]

Use backtracking to try to find all solutions of a suokdu grid. Returns Nothing if there is no solution.

nextPos :: Pos -> Maybe Pos

Get the next position in a grid (row-column order), if there is one.

   nextPos (0, 4) == Just (0, 5)                                                                                                        
   nextPos (2, 8) == Just (3, 0)                                                                                                        
   nextPos (8, 8) == Nothing                                                                                                            

nextBlank :: Sudoku -> Pos -> Maybe Pos

Get the next empty position in a grid, if there is one.