Sudoku.Solve.Basic

Description

Basic techniques for solving Sudoku puzzles.

Possible entries of a cell

data OneOrMany a

A list-like datatype that can contain either one or many elements

Constructors

One a

Many [a]

type Options = OneOrMany Int

Options represents our current knowledge about the potential values of a particular cell in a grid.

Utility functions for Options

countOptions :: Int -> [Options] -> Int

Count the occurrences of an Int in a list of Options

toList :: Options -> [Int]

Convert an Options object to a list of integers

makeOptions :: [Int] -> Options

Convert a list of integers to an Options object

lengthOptions :: Options -> Int

Get the length of an Options object.

deleteOption :: Int -> Options -> Options

Delete an option from an Options object.

Sudoku grid plus current knowledge

type SudokuWithOptions = Grid Options

SudokuWithOptions represents a partially complete sudoku grid, and the knowledge we have about the potential values of each cell

Utility functions for SudokuWithOptions

setOptions :: SudokuWithOptions -> Pos -> [Int] -> SudokuWithOptions

Set the options in a cell of the grid

optionsToNormal :: SudokuWithOptions -> Sudoku

Convert from SudokuWithOptions to Sudoku, only taking cells which have only 1 option

applyUntilStatic :: (SudokuWithOptions -> SudokuWithOptions) -> SudokuWithOptions -> SudokuWithOptions

Given a function that transforms a SudokuWithOptions and a starting grid, repeatedly apply the function until it converges to a fixed point.

Basic solving techniques

genInitialOptions :: Sudoku -> SudokuWithOptions

Convert from Sudoku to SudokuWithOptions, eliminating obviously impossible options.

eliminateOptions :: SudokuWithOptions -> SudokuWithOptions

eliminate options using a basic method: if a cell at position pos has only one option x, delete all occurrences of x from cells in the same row, column or box as pos

eliminateOptionsRepeatedly :: SudokuWithOptions -> SudokuWithOptions

Repeatedly apply eliminateOptions until there are no changes.

scanRows :: SudokuWithOptions -> SudokuWithOptions

Scan each row to see if there is only 1 place where a value x could be - if there is, place it there

scanCols :: SudokuWithOptions -> SudokuWithOptions

Scan each column to see if there is only 1 place where a value x could be - if there is, place it there

scanBoxes :: SudokuWithOptions -> SudokuWithOptions

Scan each box to see if there is only 1 place where a value x could be - if there is, place it there

scan :: SudokuWithOptions -> SudokuWithOptions

Apply scanRows, scanCols and scanBoxes

scanRepeatedly :: SudokuWithOptions -> SudokuWithOptions

Repeatedly apply scan until convergence.

improve :: Sudoku -> Sudoku

Improve a Sudoku grid as much as possible using the techniques defined in mBasic.