Sudoku.Types

Description

Sudoku.Types defines the datatypes used by the library and the solver.

Grid and position types

newtype Grid a

Abstract type representing a 2-dimensional grid.

Constructors

Grid [[a]]

type Sudoku = Grid Int

Sudoku is a 2D 9x9 list representing the contents of a partially- or totally-filled sudoku grid. A blank space is represented by a 0.

type Pos = (Int, Int)

Pos is an (Int, Int) tuple representing the position of a cell in a sudoku grid.

blank :: Sudoku

a blank sudoku grid

Functions on grids

(!!!) :: Grid a -> Pos -> a infixl 9

get the cell value at a position in a grid.

getRow :: Sudoku -> Int -> [Int]

get a row of the grid

getCol :: Sudoku -> Int -> [Int]

get a column of the grid

getBoxCoords :: Int -> Grid Pos

get the coordinates of a box (ordered right to left, top to bottom) as a Grid

getBoxCoordsFlat :: Int -> [Pos]

get the coordinates of a box (ordered right to left, top to bottom) as a Grid

getBox :: Grid a -> Int -> Grid a

get a box of the grid as a Grid

getBoxFlat :: Grid a -> Int -> [a]

get a box of the grid as a list

getBoxFromCoord :: Pos -> Int

get the index of the box a cell belongs to

allSquares :: Grid Pos

all cell positions as a 2D Grid

allSquaresFlat :: [Pos]

all cell positions as a list

mapRows :: ([a] -> [b]) -> Grid a -> Grid b

map over the rows of a Grid

transposeGrid :: Grid a -> Grid a

transpose a grid using transpose from Data.List

place :: Grid a -> Pos -> a -> Grid a

set an entry of a 2D grid

similar :: Pos -> Pos -> Bool

check if two positions are "similar", i.e. in the same rowboxcol but NOT the same

replaceValues :: (Show a, Eq a) => [(a, a)] -> Grid a -> Grid a

given a grid and a list representing a map, replace all values of the grid according to the map