Sudoku.Util

Description

A collection of utility functions for working with Sudoku grids.

Printing Sudoku grids

showSudoku :: Sudoku -> String

convert a grid to a simple string representation - blanks are represented as .

showSudokuNice :: Sudoku -> String

convert a grid to a pretty string representation

printSudoku :: Sudoku -> IO ()

print a sudoku grid in its simple string representation - see showSudoku

printSudokuNice :: Sudoku -> IO ()

print a sudoku grid in its pretty string representation - see showSudokuNice

Parsing Sudoku grids

parseSudoku :: String -> Sudoku

parse a sudoku grid from a string

readFromFile :: FilePath -> IO Sudoku

read a file and parse the sudoku grid

readFromCSV :: FilePath -> IO [(Sudoku, Sudoku)]

Parse puzzles and solutions a CSV file in the format in this kaggle dataset