Sudoku.Validity

Description

Functions to check the validity and correctness of a Sudoku grid.

Grid predicates

isValid :: Sudoku -> Bool

check that a grid is valid

isRightSize :: Sudoku -> Bool

check that a grid has the right dimensions

isValidRows :: Sudoku -> Bool

check that a grid has no rows with duplicate elements

isValidCols :: Sudoku -> Bool

check that a grid has no cols with duplicate elements

isValidBoxes :: Sudoku -> Bool

check that a grid has no boxes with duplicate elements

Solution predicates

isSolved :: Sudoku -> Bool

check that a grid is solved, i.e. complete and valid