SudokuAbstract

Documentation

type Index = Int

Index represents a group, column or a row

type Index2D = (Index, Index)

Index represents a square in 2D

nextIndex :: (Num t1, Num t, Eq t) => (t, t1) -> (t, t1)

nextIndex returns next Index2D

class (Show t, Eq t) => SudokuValue t where

SudokuValue is a class of types representing sudoku values (1-9).

Minimal complete definition

fromInt, getAll

Methods

fromInt :: Int -> t

getAll :: [t]

fromInt :: SudokuValue t => Int -> t

getAll :: SudokuValue t => [t]

class SudokuSquare t where

SudokuSquare is a class of types representing a square in susoku (it either has a value, or it is empty).

Minimal complete definition

value, isKnown, known, unknown

Methods

value :: SudokuValue value => t value -> Maybe value

Returns value wrapped in Maybe. Returns Nothing if value is not known.

isKnown :: SudokuValue value => t value -> Bool

Returns true if value is known, otherwise false

known :: SudokuValue value => value -> t value

Takes a SudokuValue and returns a square with this value

unknown :: SudokuValue value => t value

Returns unknown square

value :: (SudokuSquare t, SudokuValue value) => t value -> Maybe value

Returns value wrapped in Maybe. Returns Nothing if value is not known.

isKnown :: (SudokuSquare t, SudokuValue value) => t value -> Bool

Returns true if value is known, otherwise false

known :: (SudokuSquare t, SudokuValue value) => value -> t value

Takes a SudokuValue and returns a square with this value

unknown :: (SudokuSquare t, SudokuValue value) => t value

Returns unknown square

class SudokuBoard t where

SudokuBoard is a class of types representing sudoku board.

Minimal complete definition

element, replace, getRow, getColumn, getGroup, fromIntList

Methods

element :: (SudokuSquare square, SudokuValue val) => t (square val) -> Index2D -> square val

Returns square from given position.

replace :: (SudokuSquare square, SudokuValue val) => t (square val) -> Index2D -> square val -> t (square val)

Replace square at given position with a new square

getRow :: (SudokuSquare square, SudokuValue val) => t (square val) -> Index -> [val]

Returns list of values from given row

getColumn :: (SudokuSquare square, SudokuValue val) => t (square val) -> Index -> [val]

Returns list of values from given column

getGroup :: (SudokuSquare square, SudokuValue val) => t (square val) -> Index -> [val]

Returns list of values from given group (square 3x3)

fromIntList :: (SudokuSquare square, SudokuValue val) => [Int] -> t (square val)

Takes a list of values in int representation (0-9) and returns a board with those values

element :: (SudokuBoard t, SudokuSquare square, SudokuValue val) => t (square val) -> Index2D -> square val

Returns square from given position.

replace :: (SudokuBoard t, SudokuSquare square, SudokuValue val) => t (square val) -> Index2D -> square val -> t (square val)

Replace square at given position with a new square

getRow :: (SudokuBoard t, SudokuSquare square, SudokuValue val) => t (square val) -> Index -> [val]

Returns list of values from given row

getColumn :: (SudokuBoard t, SudokuSquare square, SudokuValue val) => t (square val) -> Index -> [val]

Returns list of values from given column

getGroup :: (SudokuBoard t, SudokuSquare square, SudokuValue val) => t (square val) -> Index -> [val]

Returns list of values from given group (square 3x3)

fromIntList :: (SudokuBoard t, SudokuSquare square, SudokuValue val) => [Int] -> t (square val)

Takes a list of values in int representation (0-9) and returns a board with those values

squaresValues :: (SudokuSquare s, SudokuValue v) => [s v] -> [v]

squaresValues takes a list of squares and returns their values.

valKnown :: (SudokuValue value, SudokuSquare t, SudokuBoard t1) => t1 (t value) -> Index2D -> Bool

Helper function, returns true if square at given position is known

possibilities :: (SudokuBoard b, SudokuSquare s, SudokuValue v) => b (s v) -> Index2D -> [v]

possibilities takes a board, index2d and returns possible values for the square

readBoard :: (SudokuValue v, SudokuSquare s, SudokuBoard b) => String -> Maybe (b (s v))

Takes a string and returns a filled board